# openapi.yaml
openapi: 3.1.0
info:
  title: CrediGraph API
  description: API for querying domain credibility scores.
  version: "0.4.1"
  contact:
    name: Complex Data Lab
  license:
    name: MIT

servers:
  - url: https://credi-net-credinet.hf.space
    description: Production server

paths:
  /:
    get:
      operationId: root
      summary: Service root endpoint
      tags: [System]
      responses:
        '200':
          description: Root response
          content:
            application/json:
              schema:
                type: object
                required: [status, api_version, docs]
                properties:
                  status:
                    type: string
                    example: ok
                  api_version:
                    type: string
                    example: "0.4.0"
                  docs:
                    type: string
                    example: "/docs"

  /health:
    get:
      operationId: health
      summary: Health check endpoint
      tags: [System]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required: [status, api_version]
                properties:
                  status:
                    type: string
                    enum: [ok, degraded]
                    example: ok
                  api_version:
                    type: string
                    example: "0.4.0"

  /metadata:
    get:
      operationId: metadata
      summary: API and score metadata
      tags: [System]
      responses:
        '200':
          description: Metadata response
          content:
            application/json:
              schema:
                type: object
                required: [api_version, data_cutoff_month, method, score_sources]
                properties:
                  api_version:
                    type: string
                    example: "0.4.0"
                  data_cutoff_month:
                    type: string
                    example: "2024-12"
                  method:
                    type: string
                    example: "Content + topology-based"
                  score_sources:
                    type: object
                    required: [regression, binary]
                    properties:
                      regression:
                        type: string
                        format: uri
                      binary:
                        type: string
                        format: uri

  /by_domain:
    post:
      operationId: queryDomain
      summary: Public single-domain query
      description: Returns model-derived credibility values for one domain.
      tags: [Queries]
      parameters:
        - $ref: '#/components/parameters/DomainQuery'
      responses:
        '200':
          description: Successful query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoredDomainResult'
        '400':
          description: Bad request (invalid domain)
        '404':
          description: Domain not found in model outputs

  /by_domains:
    post:
      operationId: queryDomains
      summary: Public batch query
      description: Returns model-derived credibility values for multiple domains.
      tags: [Queries]
      requestBody:
        $ref: '#/components/requestBodies/DomainList'
      responses:
        '200':
          description: Successful query
          content:
            application/json:
              schema:
                type: array
                items:
                  anyOf:
                    - $ref: '#/components/schemas/ScoredDomainResult'
                    - $ref: '#/components/schemas/DomainNotFoundResult'
        '400':
          description: Bad request (invalid domain)

  /by_domain_cts:
    post:
      operationId: queryDomainContinuous
      summary: Internal single-domain continuous prediction
      description: Returns continuous (regression) credibility predictions from model.
      tags: [Internal]
      parameters:
        - $ref: '#/components/parameters/DomainQuery'
      responses:
        '200':
          description: Successful query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoredDomainResult'
        '400':
          description: Bad request (invalid domain)
        '404':
          description: Domain not found in model outputs

  /by_domains_cts:
    post:
      operationId: queryDomainsContinuous
      summary: Internal batch continuous prediction
      description: Returns continuous (regression) credibility predictions from model.
      tags: [Internal]
      requestBody:
        $ref: '#/components/requestBodies/DomainList'
      responses:
        '200':
          description: Successful query
          content:
            application/json:
              schema:
                type: array
                items:
                  anyOf:
                    - $ref: '#/components/schemas/ScoredDomainResult'
                    - $ref: '#/components/schemas/DomainNotFoundResult'
        '400':
          description: Bad request (invalid domain)

  /by_domain_dr:
    post:
      operationId: queryDomainDomainRel
      summary: Internal single-domain DomainRel query
      description: Returns binary credibility labels from DomainRel ground-truth.
      tags: [Internal]
      parameters:
        - $ref: '#/components/parameters/DomainQuery'
      responses:
        '200':
          description: Successful query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoredDomainResult'
        '400':
          description: Bad request (invalid domain)
        '404':
          description: Domain not found in DomainRel labels

  /by_domains_dr:
    post:
      operationId: queryDomainsDomainRel
      summary: Internal batch DomainRel query
      description: Returns binary credibility labels from DomainRel ground-truth.
      tags: [Internal]
      requestBody:
        $ref: '#/components/requestBodies/DomainList'
      responses:
        '200':
          description: Successful query
          content:
            application/json:
              schema:
                type: array
                items:
                  anyOf:
                    - $ref: '#/components/schemas/ScoredDomainResult'
                    - $ref: '#/components/schemas/DomainNotFoundResult'
        '400':
          description: Bad request (invalid domain)

  /by_domain_dqr:
    post:
      operationId: queryDomainDQR
      summary: Internal single-domain DQR query
      description: Returns continuous credibility labels from DQR ground-truth.
      tags: [Internal]
      parameters:
        - $ref: '#/components/parameters/DomainQuery'
      responses:
        '200':
          description: Successful query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoredDomainResult'
        '400':
          description: Bad request (invalid domain)
        '404':
          description: Domain not found in DQR labels

  /by_domains_dqr:
    post:
      operationId: queryDomainsDQR
      summary: Internal batch DQR query
      description: Returns continuous credibility labels from DQR ground-truth.
      tags: [Internal]
      requestBody:
        $ref: '#/components/requestBodies/DomainList'
      responses:
        '200':
          description: Successful query
          content:
            application/json:
              schema:
                type: array
                items:
                  anyOf:
                    - $ref: '#/components/schemas/ScoredDomainResult'
                    - $ref: '#/components/schemas/DomainNotFoundResult'
        '400':
          description: Bad request (invalid domain)

components:
  parameters:
    DomainQuery:
      name: domain
      in: query
      required: true
      description: Single domain to query.
      schema:
        type: string
        pattern: '^(?:[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?\.)+[a-z0-9-]{2,}$'
        example: apnews.com

  requestBodies:
    DomainList:
      required: true
      content:
        application/json:
          schema:
            type: array
            items:
              type: string
              minLength: 1
              pattern: '^(?:[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?\.)+[a-z0-9-]{2,}$'
            example: [apnews.com, cnn.com]

  schemas:
    ScoredDomainResult:
      type: object
      required: [domain]
      properties:
        domain:
          type: string
          example: apnews.com
        credibility_level:
          type: number
          format: float
          description: Continuous credibility score.
          example: 0.85
        credible:
          type: boolean
          description: Binary credibility score.
          example: true

    DomainNotFoundResult:
      type: object
      required: [domain]
      properties:
        domain:
          type: string
          description: Queried domain not found in the relevant source.
          example: unknown.example

tags:
  - name: Queries
    description: Public model-based query endpoints
  - name: Internal
    description: Internal query endpoints (model continuous, DomainRel binary, DQR regression)
  - name: System
    description: Service status and metadata endpoints

x-versioning:
  currentVersion: "0.4.0"
  deprecationPolicy: |
    API versions are deprecated 6 months after a new version is released.
  clientVersionTracking: |
    Server tracks client versions via User-Agent header.
    Clients should include: User-Agent: credigraph/X.Y.Z